+++ ouds_theme_contract/lib/config/component/ouds_button_config.dart
@@ -11,10 +11,15 @@
  * //
  */
 
+enum OudsButtonConfigBorder {
+  radiusDefault,
+  radiusRounded;
+}
+
 class OudsButtonConfig {
-  final bool? rounded;
+  final OudsButtonConfigBorder rounded;
 
   const OudsButtonConfig({
-    this.rounded = true,
+    this.rounded = OudsButtonConfigBorder.radiusDefault,
   });
 }